Consolidate cache updates into own directory
