+++ src/components/graphql/cache/index.js
@@ -1,0 +1,7 @@
+import updateAfterPersonChange from './updateAfterPersonChange';
+import updateAfterPersonDelete from './updateAfterPersonDelete';
+
+export {
+    updateAfterPersonChange,
+    updateAfterPersonDelete
+};
